made progress towards main page
