+++ ChordCraft/ContentView.swift
@@ -5,56 +5,29 @@
 //
 
 import SwiftUI
-import SwiftData
 
 struct ContentView: View {
-    @Environment(\.modelContext) private var modelContext
-    @Query private var items: [Item]
+    @State var currentOption = 0     // State variable to track the current selected option
     
-    //variables to store input
-    @State private var buttonText = "Add"
-    @State private var songName: String = ""
-    @State private var songG: String = ""
-    var body: some View {
-            VStack {
-                // text boxes to take input
-                Text("ChordCraft V0.02")
-                    .font(.title)
-                    .padding()
-                TextField("enter song Name",text: $songName)
-                TextField("Enter song genre", text: $songG)
-                    .padding()
-                
-                // button logic
-                Button(action: {
-                    if(addItem()){
-                        self.buttonText = "Song added" // Update buttonText when button is tapped
-                    }else{
-                        self.buttonText = "Invalid,try again"
-                    }
-                }) {
-                    Text(buttonText)
-                        .padding()
-                        .background(Color.blue)
-                        .cornerRadius(10)
-                } // end of buttton
-                Text("list of songs Stored")
-                List { ForEach(items){item in Text(item.name)}
-                }// end of list
+    let options: [Option] = [ // Options available in the list
+        .init(title: "Projects", imageName: "folder.fill"),
+        .init(title: "Settings", imageName: "gear")]
+    
+    
+    let views: [Int: AnyView] = [     // Mapping of option index to corresponding view
+            0: AnyView(MainView()),
+            1: AnyView(Text("Testing Settings Tab"))
+        ]
+        
+        var body: some View {
+            NavigationView {        // Display the list of options
+                ListView(options: options, currentSelection: currentOption)
+                if let view = views[currentOption] {        // Display the view to the current selection
+                    view
+                }
             }
-            .buttonStyle(PlainButtonStyle())
-            .padding()
-        }
-
-    private func addItem() -> Bool{
-        withAnimation {
-            let newItem = Item(n:songName, g:songG)
-            modelContext.insert(newItem)
-            return true
-        }
+            .frame(minWidth: 600, minHeight: 400)
     }
-
-
 }
 
 #Preview {
